requirements.txt
dotenv
python-dotenv
azure-cognitiveservices-speech
python-dotenv
openai
simpleaudio